requirements.txt
# Core Django
Django==5.2.8
djangorestframework==3.16.1

# Database
psycopg2-binary==2.9.9

# API
djangorestframework-simplejwt==5.3.1
django-filter==24.3
drf-spectacular==0.27.2

# Utilities
python-decouple==3.8
Pillow==10.4.0

# Reportes
reportlab==4.2.2
openpyxl==3.1.5
pandas==2.2.2

# Deployment
gunicorn==22.0.0
whitenoise==6.7.0

# Development
django-debug-toolbar==4.4.6

# Base dependencies
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2
pytz==2024.1
